+++ starlette/requests.py
@@ -5,7 +5,7 @@ from urllib.parse import unquote
 import http.cookies
 from starlette.datastructures import URL, Headers, QueryParams
 from starlette.formparsers import FormParser, MultiPartParser
-from starlette.types import Scope, Receive
+from starlette.types import Scope, Receive, Message
 
 try:
     from multipart.multipart import parse_options_header
@@ -17,11 +17,15 @@ class ClientDisconnect(Exception):
     pass
 
 
+async def empty_receive() -> Message:
+    raise RuntimeError("Receive channel has not been made available")
+
+
 class Request(Mapping):
     def __init__(self, scope: Scope, receive: Receive = None) -> None:
         assert scope["type"] == "http"
         self._scope = scope
-        self._receive = receive
+        self._receive = empty_receive if receive is None else receive
         self._stream_consumed = False
 
     def __getitem__(self, key: str) -> str:
@@ -72,6 +76,10 @@ class Request(Mapping):
             self._cookies = cookies
         return self._cookies
 
+    @property
+    def receive(self) -> Receive:
+        return self._receive
+
     def url_for(self, name: str, **path_params: typing.Any) -> URL:
         router = self._scope["router"]
         url = router.url_path_for(name, **path_params)
@@ -84,9 +92,6 @@ class Request(Mapping):
 
         if self._stream_consumed:
             raise RuntimeError("Stream consumed")
-
-        if self._receive is None:
-            raise RuntimeError("Receive channel has not been made available")
 
         self._stream_consumed = True
         while True: